Implements debug zone {break: true}
This implements a new option in the debug zone
(`require("can-zone/debug");`): {break: true}

This option enables a breakpoint to occur *after* the Zone has timed
out. A `debugger;` is placed within a function that you can easily step
into to see the code that is preventing the Zone to complete.

Closes #148